requirements.txt
streamlit
PyPDF2
langchain
langchain-core
langchain-text-splitters
langchain-google-genai
langchain-chroma
chromadb
